serious progress.
datasize still borked.

+++ src/micronix/include/alloc.h
@@ -1,14 +1,15 @@
 /*
  * header file for alloc, free and realloc
  *
- * include/alloc.h
- * Changed: <2021-12-23 14:41:14 curt>
+ * /include/alloc.h
+ *
+ * Changed: <2023-07-04 11:17:40 curt>
  */
 
-# define pool _pool
+#define pool _pool
 
 struct pool {
-	UINT size;
+	unsigned short size;
 	struct pool *next;
 };
